% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv2NamesFile.R
\name{conv2NamesFile}
\alias{conv2NamesFile}
\title{Processing data with two or more heading rows}
\usage{
conv2NamesFile(fileIn)
}
\arguments{
\item{fileIn}{An input nano string data file name, which can be assoicated with an absolute path or a relative path starting from your current path.
The file itself contains two or more rows of heading, which is not usually data format for R, but we need it for nano string DLBCL90 to make
sure we have enough information}
}
\value{
A string vector containing two items, the 1st one is the regular data format name, and the 2nd one is the file name with extra sample info
}
\description{
This is to separate input data containing two or more heading rows into regular data format table and one info table
}
\details{
This function reads in full path csv file with two or more rows of heading info, then write out a usual one row heading data table and an 
 info table contain all heading info in the same path of the input file.
}
\author{
Aixiang Jiang
}
\keyword{nano}
\keyword{string}
