% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLBCL90calls.R
\name{DLBCL90calls}
\alias{DLBCL90calls}
\title{A wrap-up function to get all calls: PMBL, DLBCL and DHITsig with two different data formats}
\usage{
DLBCL90calls(DLBCL90_File, outfileName, nHeading = 2, geomeanCut = 60)
}
\arguments{
\item{DLBCL90_File}{A nano string data file name, which can be assoicated with an absolute path or a relative path starting from your current path,
it can be either raw data as a zip file name, or a combined data file in csv format.}

\item{outfileName}{An output file name, which can be assoicated with an absolute path or a relative path starting from your current path}

\item{nHeading}{A integer of number of heading rows, default is 2 since most recently files are with two rows of heading info, however, 
I only check if this is 1 or not}

\item{geomeanCut}{A QC (quality control) cutoff for house keeping genes' mean, default is 60, 
which is close to 2**6 = 64. Here, 6 is the QC cutoff for PMBL/DLBCL calls that is done in log2 scale}
}
\value{
A data frame with PMBL + DLBCL + DHIT sig calls and their related values. At the same time, this function writes result table 
  into the same path of input data, also writes one row heading data table as well heading info table if input file contains two heading rows.
}
\description{
This is the wrap up function to get all done and for input data with either 1 and more heading rows.
}
\details{
This includes: nano string data re-formatring when necessary, nano string pre-process, PMBL + DLBCL + DHIT sig calls, and combination
}
\author{
Aixiang Jiang
}
\keyword{DLBCL90}
\keyword{nano-string}
