% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lymph3Cx-to-DLBCL90.R
\name{Lymph3CX.to.DLBCL90}
\alias{Lymph3CX.to.DLBCL90}
\title{Lymph3CX for a nano string data}
\usage{
Lymph3CX.to.DLBCL90(data, model.table, include.shift = T)
}
\arguments{
\item{data}{a nano string data set in data frame or data matrix format, rows are for genes (symbols), and columns are for samples
which should not be log transformed since log2 transformation is included in the function}

\item{model.table}{a known model table with fixed setting}

\item{include.shift}{a logic variable to indicate that if calibration should be included in the function.
Its default value is TRUE, calibration is done with adding baseline shift from model table; if FALSE is chosen, however, 
the calibration is reversed.}
}
\value{
A data frame with normval, PMBLval, PMBLp, PMBLcall, DLBCLval, DLBCLp, DLBCLcall, Totalcall
}
\description{
This function is to get PMBL calls (DLBCL,PMBL and Unclear PMBL/GCB) and DLBCL calls (COO, cell of origin, ABC，GCB and Unclass DLBCL)
 for a nano string data set
}
\details{
Both PMBL and DLBCL classification scores are based on LPS algorithm from Wright (2003), 
 and their clustering probability calculation is Empiracal Bayes' based.
}
\references{
Wright G, Tan B, Rosenwald A, Hurt EH, Wiestner A, Staudt LM. A gene expression-based method
to diagnose clinically distinct subgroups of diffuse large B cell lymphoma. Proc Natl Acad Sci U S
A. 2003 Aug 19;100(17):9991-6.
}
\author{
Stacy Hung (function), Aixiang Jiang (function desciption)
}
\keyword{DLBCL}
\keyword{LPS}
\keyword{PMBL}
