% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeX.R
\name{removeX}
\alias{removeX}
\title{This is a simple function to remove "X" at the beginning of a string for a  vector of strings}
\usage{
removeX(namesIn)
}
\arguments{
\item{namesIn}{a vector of string}
}
\value{
a string vector
}
\description{
This is a simple function to remove "X" at the beginning of a string for a  vector of strings
}
\author{
Aixiang Jiang
}
